Drop the mask on the GemmKernel
Write masked_kernel as a generic operation
